Fix product import without categories and tax categories
