+++ src/main/java/com/commercetools/dataimport/LocalizedField.java
@@ -1,0 +1,25 @@
+package com.commercetools.dataimport;
+
+import io.sphere.sdk.models.LocalizedString;
+import io.sphere.sdk.models.LocalizedStringEntry;
+import lombok.Data;
+
+import java.util.Locale;
+import java.util.stream.Stream;
+
+import static org.apache.commons.lang3.StringUtils.isEmpty;
+
+@Data
+public class LocalizedField {
+
+    private String de;
+    private String en;
+
+    public LocalizedString toLocalizedString() {
+        return Stream.of(
+                LocalizedStringEntry.of(Locale.GERMAN, getDe()),
+                LocalizedStringEntry.of(Locale.ENGLISH, getEn()))
+                .filter(entry -> !isEmpty(entry.getValue()))
+                .collect(LocalizedString.streamCollector());
+    }
+}
