+++ src/main/java/com/commercetools/dataimport/categories/CategoryCsvEntry.java
@@ -1,5 +1,6 @@
 package com.commercetools.dataimport.categories;
 
+import com.commercetools.dataimport.LocalizedField;
 import lombok.Data;
 
 @Data
